Add notification for task when on time

+++ app/src/main/java/com/trungdunghoang125/mytasks/view/adapter/ItemClick.java
@@ -3,7 +3,7 @@ package com.trungdunghoang125.mytasks.view.adapter;
 import com.trungdunghoang125.mytasks.model.Task;
 
 public interface ItemClick {
-    public void onItemClick(Long pos);
+    public void onItemClick(int pos);
 
     public void onCbTaskDoneClick(Task task, Boolean state);
